+++ binx/collection.py
@@ -4,6 +4,9 @@
 import abc
 import pandas as pd
 import numpy as np
+import copy
+import uuid
+
 from marshmallow import Schema, post_load, fields
 from marshmallow.exceptions import ValidationError
 
@@ -178,7 +181,7 @@ class BaseCollection(AbstractCollection):
     def __init__(self):
         self._data = []
         self._serializer = self.serializer_class(internal=self.__class__.internal_class, strict=True)
-
+        self.__collection_id = uuid.uuid4().hex
 
     @classmethod
     def get_fully_qualified_class_path(cls):
@@ -214,6 +217,10 @@ class BaseCollection(AbstractCollection):
         """
         return self.__class__.internal_class
 
+    @property
+    def collection_id(self):
+        return self.__collection_id
+
 
     def __iter__(self):
         self._idx = 0
@@ -245,7 +252,7 @@ class BaseCollection(AbstractCollection):
             raise TypeError('Only Collections of the same class can be concatenated')
 
     @classmethod
-    def _resolve_adapter_chain(cls, input_collection, **adapter_context):
+    def _resolve_adapter_chain(cls, input_collection, accumulate, **adapter_context):
         """ attempts to resolve the adapter chain using the current class as the target and
         input as the starting class. The adapter context accumulates over each call and ensures that
         kwargs needed for certain adapter calls are guaranteed to make it to the correct adapter.
@@ -261,14 +268,18 @@ class BaseCollection(AbstractCollection):
         current_input = input_collection  # NOTE this is an instance with data to be transformed.. not a class
         adapter_output = None
 
-        for adapter_class in adapters:
-            #NOTE custom error handling must be added here
+        for i, adapter_class in enumerate(adapters):
+            # if accumulate make a new key in the current context for the current collection the collection name in the registry
+            if accumulate and i > 0:
+                coll_id = current_input.__class__.__name__
+                current_context[coll_id] = copy.copy(current_input)
+
             current_adapter = adapter_class() # for each adapter class we push the input_collection and a context
             adapter_output = current_adapter(current_input, **current_context) # adapt data to the next type of collection
             current_context = {**current_context, **adapter_output.context} # NOTE this is will fail on py3.4
             current_input = adapter_output.collection
 
-        adapter_output._context = current_context
+        adapter_output._context = current_context # set final context
         return adapter_output
 
     def _dataframe_with_dtypes(self, data):
@@ -286,7 +297,7 @@ class BaseCollection(AbstractCollection):
                 df_data[col] = pd.Series(col_data[col], dtype=dtype)
             except KeyError as err:
                 l.warning('Creating df without non-required field {}'.format(col))
-                pass 
+                pass
 
         df = pd.DataFrame(df_data)
         df = dfutil.df_none_to_nan(df)
@@ -347,7 +358,7 @@ class BaseCollection(AbstractCollection):
 
 
     @classmethod
-    def adapt(cls, input_collection, **adapter_context):
+    def adapt(cls, input_collection, accumulate=False, **adapter_context):
         """ Attempts to adapt the input collection instance into a collection of this type by
         resolving the adapter chain for the input collection. Any kwargs passed in are handed over to the resolver.
         colla = CollectionA()
@@ -357,10 +368,13 @@ class BaseCollection(AbstractCollection):
 
         This method returns a new instance of the adapted class (the caller)
         """
+
         if not issubclass(input_collection.__class__, BaseCollection): #check if its a Collection or raise TypeError
             raise TypeError('The input to adapt must be a Collection')
+
         try:
-            adapted = cls._resolve_adapter_chain(input_collection, **adapter_context) # attempt to resolve the adapter chain
+            adapted = cls._resolve_adapter_chain(input_collection, accumulate, **adapter_context) # attempt to resolve the adapter chain
+
         except Exception as err:
             raise AdapterChainError('An error occured within the adapter chain') from err